Split app.hh into two pairs of header/cpp files
